Copy XPI to test dir
This did not work until the --rev=5 option on the 'hg clone' command was removed. I don't think 5 is a valid argument for this option, as previously, there was no ./testing/profiles/ directory! This takes 3.5 minutes in total (1.5 for the initial clone and 2 to download updates). This is not ideal, and there should be more investigation into how to obtain a shallow clone of Firefox.

Added applications.gecko.id to manifest.json, which is required at least for the sample mochitest `./mach mochitest dom/security/test/cors/test_CrossSiteXHR.html` (otherwise an error related to a temporary add-on id is logged in the console, and the add-on is not installed in about:addons).