Update hg
The default version of hg on the docker image is v4. If I try to clone the Mozilla version-controls-tool hg repo to install the firefoxtree and push-to-try hg extensions (the latter is required, the former is recommended for pushing to try) and create an .hgrc pointing to those extensions' file paths in the version control tools repo, I get this error:
```
*** failed to import extension firefoxtree from ~/version-control-tools/hgext/firefoxtree: 'module' object has no attribute 'command'
*** failed to import extension push-to-try from ~/version-control-tools/hgext/push-to-try: 'module' object has no attribute 'command'
```
Updating hg to the latest version fixes this error.