Combine PR and push to master tasks
The intention is that they run the same set of tests.

Note: Taskcluster yml $if statement operators [don't short-circuit](https://github.com/taskcluster/json-e/issues/244) yet, so I can't combine the checks into something like this:
```yaml
- $if: '(tasks_for == "github-pull-request" && event["action"] in ["opened", "reopened", "synchronize"]) ||
  (tasks_for == "github-push" && event.ref == "refs/heads/master")'
  then:
```
Because if it's a pull request, event.ref doesn't exist and it throws an error.